chat/requirements.txt
langchain>=0.3.0,<0.4.0
langchain-community>=0.3.0,<0.4.0
langchain-core>=0.3.0,<0.4.0
langchain-openai>=0.2.0
langgraph>=0.2.0
openai>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.23
pydantic>=2.5.0
python-dotenv>=1.0.0
aiofiles>=23.2.0
jinja2>=3.1.2
python-multipart>=0.0.6
python-frontmatter>=1.0.0
